Major patch - new working offline back-end for Treasure Hunter!
Loads of classes, and functionality. What is required of the robot
author is to implement method getNextMove() in
MyN00bTreasureHunterGameEngine.java.

Essentially, the robot should make a decision on whether to turn left,
turn right or move forth. When a player steps on a treasure, he collects
it. Game ends when all treasures are picked up from the ground!

New classes include:
	* The Tile class (which keeps the state of each Tile). Here you
	  can see if the tile is a wall, treasure, any of the players or
	  empty.
	* Orientation enum, to know in what direction a player is
	  standing - (up, down, left, right, with directions in cartesian
	  coordinates).
	* TreasureHunterAction, the implementation of the Action
	  interface (written in an attempt to generify the back-end a
	  bit)
	* A helper TreasureHunter class to get you started with how the
	  frameworks expects you to respond.

I can't unfortunately in this state come up with more stuff that
separates this version from the latest commit - and I'm sure there are
some nasty bugs, and the code itself isn't the prettiest. Next up is
setting up the server-side of it all, to make sure two AIs can compete
against each other. Note to self: Update README